pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DescriptorDOS"
version = "0.2.4"
description = "DDOS: Score-matching the descriptor density of states"
authors = [
    { name="TD Swinburne", email="thomas.swinburne@cnrs.fr" }
]
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.21",
    "mpi4py>=3.1",
    "matplotlib>=3.4",
    "pyyaml>=6.0.2",
    "lammps>=20240829"
]

[project.urls]
Homepage = "https://github.com/tomswinburne/DescriptorDOS"
Documentation = "https://github.com/tomswinburne/DescriptorDOS"
Source = "https://github.com/tomswinburne/DescriptorDOS"

[tool.setuptools]
packages = ["DescriptorDOS"]